modularize project
add module description so it can be used in moularized projects without
warnings.
