+++ src/main/java/com/github/chaosfirebolt/converter/parser/ParserContainer.java
@@ -1,7 +1,6 @@
 package com.github.chaosfirebolt.converter.parser;
 
 import com.github.chaosfirebolt.converter.constants.IntegerType;
-import com.github.chaosfirebolt.converter.parser.impl.AbstractParser;
 
 import java.util.HashMap;
 import java.util.Map;
