+++ src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module com.github.chaosfirebolt.converter {
+  exports com.github.chaosfirebolt.converter;
+  exports com.github.chaosfirebolt.converter.constants;
+  exports com.github.chaosfirebolt.converter.exc;
+  exports com.github.chaosfirebolt.converter.parser;
+  exports com.github.chaosfirebolt.converter.util;
+}
